Search for an interface and send a UDP packet
The device `em0` or `genet0` is hard coded for use on most Intel machines and a Raspberry Pi4. The loading of the module creates and sends a single UDP packet.

The source MAC is of the found for the interface. The destination MAC and IPv4 address is for 224.0.0.1. This makes it easier to find.

This implementation uses `if_hwassist` bits and can use hardware acceleration of checksum compute if available. In case of software acceleration, the optimised routines in the FreeBSD kernel are used.

Issue: OSDEV-137